Add last modifications on Trace model.

+++ Time4Sys/plugins/org.polarsys.time4sys.marte.nfp/src/org/polarsys/time4sys/marte/nfp/impl/LongDurationImpl.java
@@ -60,7 +60,7 @@ public class LongDurationImpl extends DurationImpl {
 	}
 
 	public LongDurationImpl(final long durationInPicoseconds, final TimeUnitKind unit) {
-		super(durationInPicoseconds / asLongPicoSecond(unit), unit);
+		super(((double)durationInPicoseconds) / ((double)asLongPicoSecond(unit)), unit);
 		this.picoseconds  = durationInPicoseconds;
 		
 	}
